backend/requirements.txt
asgiref==3.7.2
axios==0.4.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
django-cors-headers==4.3.0
django-faker==0.2
django-filter==23.3
django-jazzmin==2.6.0
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
fake-factory==9999.9.9
Faker==20.0.3
idna==3.4
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
Pillow==10.1.0
psycopg2-binary==2.9.9
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rich==13.6.0
six==1.16.0
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.8.0
urllib3==2.1.0
